CB-13484 SdxRepair mocktest refactors. Splitting basic SDX test from repair use cases. Removed unnecessary env creation from testcases
